move struct to runtime
